add userId authorization

+++ Services/SavingService.cs
@@ -3,6 +3,7 @@ using HouseholdManagerApi.DTOs;
 using HouseholdManagerApi.Interfaces.Repositories;
 using HouseholdManagerApi.Interfaces.Services;
 using HouseholdManagerApi.Models;
+using Microsoft.EntityFrameworkCore;
 
 namespace HouseholdManagerApi.Services
 {
@@ -15,33 +16,49 @@ namespace HouseholdManagerApi.Services
             this.mapper = mapper;
             this.savingRepository = savingRepository;
                 }
-        public async Task<SavingDTO> Create(SavingDTO entity)
+        public async Task<SavingDTO> Create(SavingDTO dto, string userId)
         {
-            var result = await this.savingRepository.Create(this.mapper.Map<Saving>(entity));
+            var entity = this.mapper.Map<Saving>(dto);
+            entity.UserId = userId;
+            var result = await this.savingRepository.Create(entity);
 
             return this.mapper.Map<SavingDTO>(result);
         }
 
-        public async Task Delete(int id)
+        public async Task Delete(int id, string userId)
         {
+            var item = await this.GetById(id, userId);
+
+            if (item == null)
+            {
+                return;
+            }
+
             await this.savingRepository.Delete(id);
         }
 
-        public async Task<IEnumerable<SavingDTO>> GetAll()
+        public async Task<IEnumerable<SavingDTO>> GetAll(string userId)
         {
-            var result = this.savingRepository.GetAll();
+            var result = await this.savingRepository.GetAll().Where(s => s.UserId == userId).ToListAsync();
 
             return this.mapper.Map<IEnumerable<Saving>,IEnumerable<SavingDTO>>(result);
         }
 
-        public async Task<SavingDTO> GetById(int id)
+        public async Task<SavingDTO> GetById(int id, string userId)
         {
-            return this.mapper.Map<SavingDTO>(await this.savingRepository.GetById(id));
+            return this.mapper.Map<SavingDTO>(await this.savingRepository.GetAll().FirstOrDefaultAsync(s => s.UserId == userId && s.Id == id));
         }
 
-        public async Task<SavingDTO> Update(SavingDTO entity)
+        public async Task<SavingDTO> Update(SavingDTO dto, string userId)
         {
-            var res = await this.savingRepository.Update(this.mapper.Map<Saving>(entity));
+            var entity = await this.GetById(dto.Id.Value, userId);
+
+            if (entity == null)
+            {
+                return null;
+            }
+
+            var res = await this.savingRepository.Update(this.mapper.Map<Saving>(dto));
 
             return this.mapper.Map<SavingDTO>(res);
         }
